fix(revit): widen geometry-sync element ids to 64-bit so a large id can't tombstone live geometry
Revit 2024+ ElementId.Value is 64-bit and can exceed int.MaxValue in large or
long-lived models. GeometrySyncUpdater enqueued (int)id.Value and encoded a
deletion as the negated id, so a large *modified/added* element wrapped negative
and GeometrySyncPlan.Partition read it as a tombstone — the server then
soft-deleted live geometry the user had just edited. A deletion of such an id
flipped positive and was treated as a change.

Widen the whole geometry-delta id path int -> long: the queue tuple, its drain,
GeometrySyncPlan.Partition/BuildRetrySet, the handler's changed/deleted/extracted
lists and TryExtractElement, RequeueForRetry, and
PlanscapeServerClient.PostGeometryDeltaAsync(IList<long>). The server side is
already long end-to-end (FederatedElement.ElementId, the GLB node DTO); the one
remaining int on the server is the deletedIds deserialize, fixed on the
delta-atomic branch above this one.

Test: GeometrySyncPlanTests.A_large_modified_id_stays_a_change_not_a_tombstone
pins that an id past int range stays a change (not a deletion) and that a genuine
deletion of it still round-trips as a tombstone. All plan tests updated to long;
9/9 pass, plugin builds 0 errors.

NOT in scope: the changed-element *identity* still truncates via
ClashElementKey((int)el.Id.Value) -> the GLB node id. ClashElementKey is shared
clash infrastructure (ClashSession dictionaries, ClashHistory, the clash
SignalR/result DTOs and their server contract are all int), so widening it is a
separate clash-wide id migration — flagged, not bundled.

+++ StingTools/Clash/GeometrySyncPlan.cs
@@ -4,7 +4,8 @@
 // IExternalEventHandler that needs a UIApplication and a Document, so nothing in
 // it is reachable from the pure-logic test projects; but the two decisions that
 // actually govern whether a change reaches the server — how the queue encoding
-// is read, and which ids are worth retrying — are plain arithmetic over ints.
+// is read, and which ids are worth retrying — are plain arithmetic over 64-bit
+// element ids.
 using System.Collections.Generic;
 
 namespace StingTools.Core.Clash
@@ -13,8 +14,11 @@ namespace StingTools.Core.Clash
     /// The queue encoding and the retry rule for Planscape geometry sync.
     ///
     /// <para><b>The encoding.</b> <c>LiveClashUpdater.GeometrySyncQueue</c>
-    /// carries a single int per change: positive means "this element was added
-    /// or modified", negative means "this element was deleted". Packing a
+    /// carries a single 64-bit id per change: positive means "this element was
+    /// added or modified", negative means "this element was deleted". A 64-bit id
+    /// is deliberate — Revit 2024+ element ids can exceed <c>int.MaxValue</c>, and
+    /// truncating one to <c>int</c> could wrap it negative and flip an edit into a
+    /// deletion. Packing a
     /// deletion as a negated id keeps one queue and one drain, but it means the
     /// sign IS the semantics — read it backwards and every deletion becomes an
     /// attempt to tessellate an element that no longer exists, while every edit
@@ -27,13 +31,13 @@ namespace StingTools.Core.Clash
         /// ids to tombstone. Both come back POSITIVE — the sign is queue
         /// encoding, not data.
         /// </summary>
-        public static (List<int> Changed, List<int> Deleted) Partition(IEnumerable<int> drained)
+        public static (List<long> Changed, List<long> Deleted) Partition(IEnumerable<long> drained)
         {
-            var changed = new List<int>();
-            var deleted = new List<int>();
+            var changed = new List<long>();
+            var deleted = new List<long>();
             if (drained == null) return (changed, deleted);
 
-            foreach (int id in drained)
+            foreach (long id in drained)
             {
                 if (id < 0) deleted.Add(-id);
                 else if (id > 0) changed.Add(id);
@@ -57,14 +61,14 @@ namespace StingTools.Core.Clash
         /// <para>Deletions are always retryable: a tombstone needs no geometry,
         /// so there is nothing that can fail to extract.</para>
         /// </summary>
-        public static List<int> BuildRetrySet(
-            IEnumerable<int> extractedChangedIds, IEnumerable<int> deletedIds)
+        public static List<long> BuildRetrySet(
+            IEnumerable<long> extractedChangedIds, IEnumerable<long> deletedIds)
         {
-            var retry = new List<int>();
+            var retry = new List<long>();
             if (extractedChangedIds != null)
-                foreach (int id in extractedChangedIds) if (id > 0) retry.Add(id);
+                foreach (long id in extractedChangedIds) if (id > 0) retry.Add(id);
             if (deletedIds != null)
-                foreach (int id in deletedIds) if (id > 0) retry.Add(-id);
+                foreach (long id in deletedIds) if (id > 0) retry.Add(-id);
             return retry;
         }
     }
